+++ Authentication/src/main/java/org/example/presentation/view/RegisterRequestDTO.java
