--- Commons/src/main/java/org/example/errorhandling/InvalidTokenException.java
@@ -1,4 +1,0 @@
-package org.example.errorhandling;
-
-public class InvalidTokenException extends RuntimeException{
-}
