+++ ErrorHandling/src/main/java/org/example/errorhandling/customexceptions/AlreadyExistingUserException.java
@@ -1,0 +1,8 @@
+package errorhandling.customexceptions;
+
+
+public class AlreadyExistingUserException extends RuntimeException{
+    public AlreadyExistingUserException(String message, String username) {
+        super(message + username);
+    }
+}
