+++ ErrorHandling/src/main/java/org/example/errorhandling/customexceptions/InvalidCredentialsException.java
@@ -1,0 +1,7 @@
+package errorhandling.customexceptions;
+
+public class InvalidCredentialsException extends RuntimeException{
+    public InvalidCredentialsException(String message) {
+        super(message);
+    }
+}
