+++ ErrorHandling/src/main/java/org/example/errorhandling/customexceptions/InvalidTokenException.java
@@ -1,0 +1,4 @@
+package errorhandling.customexceptions;
+
+public class InvalidTokenException extends RuntimeException{
+}
