+++ ErrorHandling/src/main/java/org/example/errorhandling/customexceptions/InvalidUpdateException.java
@@ -1,0 +1,12 @@
+package errorhandling.customexceptions;
+
+import java.util.UUID;
+
+/**
+ * This exception is thrown when an update request cannot be performed
+ */
+public class InvalidUpdateException extends RuntimeException{
+    public InvalidUpdateException(String message, UUID uuid) {
+        super(message + uuid);
+    }
+}
