+++ ErrorHandling/src/main/java/org/example/errorhandling/customexceptions/InvoiceNotFoundException.java
@@ -1,0 +1,9 @@
+package errorhandling.customexceptions;
+
+import java.util.UUID;
+public class InvoiceNotFoundException extends RuntimeException{
+        public InvoiceNotFoundException(String message) {
+            super(message);
+        }
+    }
+
