+++ ErrorHandling/src/main/java/org/example/errorhandling/customexceptions/UserNotFoundException.java
@@ -1,0 +1,10 @@
+package errorhandling.customexceptions;
+
+/**
+ * This exception is thrown whenever a non-existent user is queried
+ */
+public class UserNotFoundException extends RuntimeException{
+    public UserNotFoundException(String message, String username) {
+        super(message + username);
+    }
+}
