--- InvoiceFeature2/src/main/java/org/example/business/exceptions/InvoiceNotFoundException.java
@@ -1,9 +1,0 @@
-package org.example.business.exceptions;
-
-import java.util.UUID;
-public class InvoiceNotFoundException extends RuntimeException{
-        public InvoiceNotFoundException(String message) {
-            super(message);
-        }
-    }
-
