--- PurchaseOrder/src/main/java/org/example/business/errorhandling/ErrorMessages.java
@@ -1,7 +1,0 @@
-package org.example.business.errorhandling;
-
-public class ErrorMessages {
-    public static final String ORDER_NOT_FOUND = "Could not find order identified by ";
-    public static final String INVALID_UPDATE = "It is not allowed to update the order identified by  ";
-    public static final String INVALID_VERSION = "Your resource representation may not be updated.";
-}
