--- PurchaseOrder/src/main/java/org/example/business/errorhandling/customexceptions/InvalidUpdateException.java
@@ -1,12 +1,0 @@
-package org.example.business.errorhandling.customexceptions;
-
-import java.util.UUID;
-
-/**
- * This exception is thrown when an update request cannot be performed
- */
-public class InvalidUpdateException extends RuntimeException{
-    public InvalidUpdateException(String message, UUID uuid) {
-        super(message + uuid);
-    }
-}
