--- UserAuth/src/main/java/org/example/business/errorhandling/customexceptions/AlreadyExistingUserException.java
@@ -1,10 +1,0 @@
-package org.example.business.errorhandling.customexceptions;
-
-import java.util.UUID;
-
-
-public class AlreadyExistingUserException extends RuntimeException{
-    public AlreadyExistingUserException(String message, String username) {
-        super(message + username);
-    }
-}
