--- UserAuth/src/main/java/org/example/business/errorhandling/customexceptions/UserNotFoundException.java
@@ -1,12 +1,0 @@
-package org.example.business.errorhandling.customexceptions;
-
-import java.util.UUID;
-
-/**
- * This exception is thrown whenever a non-existent user is queried
- */
-public class UserNotFoundException extends RuntimeException{
-    public UserNotFoundException(String message, String username) {
-        super(message + username);
-    }
-}
