--- core/src/main/java/org/example/filters/TestFilter.java
@@ -1,34 +1,0 @@
-package org.example.filters;
-
-import org.example.utils.TokenHandler;
-
-import javax.servlet.*;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-import java.util.Collections;
-import java.util.List;
-import java.util.Map;
-import java.util.function.Function;
-import java.util.stream.Collectors;
-
-public class TestFilter implements Filter {
-    @Override
-    public void doFilter(ServletRequest servletRequest, ServletResponse servletResponse, FilterChain filterChain) throws IOException, ServletException {
-        System.out.println("In filter method");
-        HttpServletRequest req = (HttpServletRequest) servletRequest;
-        HttpServletResponse res = (HttpServletResponse) servletResponse;
-
-        Map<String, List<String>> headersMap = Collections.list(req.getHeaderNames())
-                .stream()
-                .collect(Collectors.toMap(
-                        Function.identity(),
-                        h -> Collections.list(req.getHeaders(h))
-                ));
-
-        String token = headersMap.get("authorization").get(0);
-        TokenHandler.validateToken(token);
-
-        filterChain.doFilter(servletRequest, servletResponse);
-    }
-}
